requirements.txt
pandas
numpy
scikit-learn
pyyaml
transformers
torch
streamlit
duckdb
pyarrow
fastparquet
tqdm
